Fixed x86_64 compile errors

+++ az_src/src/org/gudy/azureus2/ui/swt/test/PrintTransferTypes.java
@@ -40,7 +40,7 @@ public class PrintTransferTypes extends ByteArrayTransfer {
 
         TransferData[] data = event.dataTypes;
         for (int i = 0; i < data.length; i++) {
-          int id = data[i].type;
+          int id = (int)data[i].type;
           String name = getNameFromId(id);
           System.out.println("Data type is " + id + " " + name);
         }
@@ -50,7 +50,7 @@ public class PrintTransferTypes extends ByteArrayTransfer {
       }
       public void drop(DropTargetEvent event) {
         System.out.println("URL dropped: " + event.data);
-        System.out.println("Data type is " + event.currentDataType.type + " " + getNameFromId(event.currentDataType.type));
+        System.out.println("Data type is " + event.currentDataType.type + " " + getNameFromId((int)event.currentDataType.type));
       }
     });
